support experiment

+++ bin/c8_driver_perf.sh
@@ -17,6 +17,7 @@
 noofrecords_default=1000
 noofthreads_default=10
 debug_default=false
+local_default=false
 result_file=data/c8_performance_runs.dat
 
 usage() {
@@ -29,6 +30,7 @@ usage: `basename ${0}` [options] [commands]
 options:
   -n | --noofrecords [ long ]            : Total no of records(partition keys to be written, there are 4 cluserting columns for every partition key); default: $noofrecords_default
   -t | --noofthreads [ integer ]         : no of threads to run; default: $noofthreads_default
+  -l | --local [ true | false ]          : use local Cassandra instead of docker $local_default
   -d | --debug [ true | false ]          : debug ; default: $debug_default
   -h | --help                            : help message
 
@@ -59,6 +61,7 @@ while getopts "${optspec}" opt; do
     n) noofrecords=${OPTARG};;
     t) noofthreads=${OPTARG};;
     d) debug=${OPTARG};;
+    l) local=${OPTARG};;
     h) usage;;
     :) usage "option -${OPTARG} requires an argument" 1;;
     \?) [ "${OPTERR}" != 1 ] || [ "${optspec:0:1}" = ":" ] && usage "unknown option -${OPTARG}" 1;;
@@ -75,6 +78,7 @@ run() {
   echo "noofthreads: $noofthreads "
   echo "debug: $debug "
   echo "driver: $driver "
+  echo "local: $local "
   echo ""
 
   ls modules/sprng_data_1_4_2/target/*-c8-perf-all.jar | grep 'c8-perf-all' &> /dev/null
@@ -91,25 +95,32 @@ run() {
      return -1
   fi
 
-  docker --version | grep 'version' &> /dev/null
-  if [ $? != 0 ]; then
-     echo "Error: Docker not installed"
-     echo "Please install Docker"
-     return -1
-  fi
+  if [ "$local" == false ] ; then
+  	  docker --version | grep 'version' &> /dev/null
+	  
+	  if [ $? != 0 ]; then
+	     echo "Error: Docker not installed"
+	     echo "Please install Docker"
+	     return -1
+	  fi
+	  
+	  docker stop -f $docker_name 2> /dev/null
+	  docker rm -f $docker_name 2> /dev/null
+	  docker run -d --name $docker_name -p 9042:9042 -p 9160:9160 -d cassandra:2.1
+	  
+	  cql_setup=`cat ./bin/c8_driver_perf_setup.cql`
+	  echo "Setting up Cassandra: $cql_setup"
+	
+	  # Check if docker ready to accept cqlsh commands then remove sleep
+	  sleep 20
+	  docker exec -it $docker_name cqlsh -e "$cql_setup"
+	  echo "Command for cqlsh: docker exec -it $docker_name cqlsh"
+  else
+      $CASSANDRA_HOME/bin/cqlsh -f /root/Documents/repo/c8driver_perf/bin/c8_driver_perf_setup.cql	  	  
+  fi	  
 
-  docker stop -f $docker_name 2> /dev/null
-  docker rm -f $docker_name 2> /dev/null
-  docker run -d --name $docker_name -p 9042:9042 -p 9160:9160 -d cassandra:2.1
-
-  cql_setup=`cat ./bin/c8_driver_perf_setup.cql`
-  echo "Setting up Cassandra: $cql_setup"
-
-  # Check if docker ready to accept cqlsh commands then remove sleep
-  sleep 20
-  docker exec -it $docker_name cqlsh -e "$cql_setup"
-  echo "Command for cqlsh: docker exec -it $docker_name cqlsh"
-
+  
+  
   mkdir -p data
 
   start_date=`date`
@@ -122,8 +133,10 @@ run() {
 
   echo "Process Started at: $start_date"
   echo "Process Ended at: $end_date"
-  docker stop $docker_name
-  docker rm $docker_name
+  if [ "$local" == false ] ; then
+	  docker stop $docker_name
+	  docker rm $docker_name
+  fi		  
 }
 
 [ $# -eq 0 ] && usage "unspecified command" 1
